langchain_llm_chatbot/requirements.txt
streamlit
langchain
langchain-core
langchain-community
google-genai
python-dotenv
pypdf
